requirements.txt
streamlit>=1.36.0
google-genai>=0.7.0